+++ app/src/main/java/com/sopt/withsuhyeon/feature/onboarding/GenderSelectScreen.kt
@@ -176,7 +176,10 @@ fun GenderSelectScreen(
             }
         }
 
-        HorizontalDivider(modifier = Modifier.height(1.dp))
+        HorizontalDivider(
+            modifier = Modifier.height(1.dp),
+            color = colors.Grey100
+        )
         Spacer(modifier = Modifier.height(16.dp))
         LargeButton(
             onClick = onButtonClick,
